# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"SentiNews Finance RAG","description":"Finance RAG system — Phase 1 (foundation).","version":"0.1.0"},"paths":{"/api/health":{"get":{"tags":["health"],"summary":"Health","description":"Reports backend liveness plus a live check of the database\nconnection and whether the pgvector extension is installed.","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/documents/upload":{"post":{"tags":["documents"],"summary":"Upload Document","operationId":"upload_document_api_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_documents_upload_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentUploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents":{"get":{"tags":["documents"],"summary":"List Documents","operationId":"list_documents_api_documents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DocumentListItem"},"type":"array","title":"Response List Documents Api Documents Get"}}}}}}},"/api/documents/{document_id}":{"get":{"tags":["documents"],"summary":"Get Document","operationId":"get_document_api_documents__document_id__get","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Document Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/retrieval/search":{"post":{"tags":["retrieval"],"summary":"Search","operationId":"search_api_retrieval_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RetrievalSearchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RetrievalSearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Root  Get"}}}}}}}},"components":{"schemas":{"Body_upload_document_api_documents_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_document_api_documents_upload_post"},"DatabaseStatus":{"properties":{"connected":{"type":"boolean","title":"Connected"},"pgvector_installed":{"type":"boolean","title":"Pgvector Installed"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["connected","pgvector_installed"],"title":"DatabaseStatus"},"DocumentDetail":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"filename":{"type":"string","title":"Filename"},"file_type":{"type":"string","title":"File Type"},"file_size":{"type":"integer","title":"File Size"},"status":{"$ref":"#/components/schemas/DocumentStatus"},"page_count":{"type":"integer","title":"Page Count"},"chunk_count":{"type":"integer","title":"Chunk Count"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"}},"type":"object","required":["id","filename","file_type","file_size","status","page_count","chunk_count","created_at","updated_at"],"title":"DocumentDetail"},"DocumentListItem":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"filename":{"type":"string","title":"Filename"},"file_type":{"type":"string","title":"File Type"},"file_size":{"type":"integer","title":"File Size"},"status":{"$ref":"#/components/schemas/DocumentStatus"},"page_count":{"type":"integer","title":"Page Count"},"chunk_count":{"type":"integer","title":"Chunk Count"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","filename","file_type","file_size","status","page_count","chunk_count","created_at","updated_at"],"title":"DocumentListItem"},"DocumentStatus":{"type":"string","enum":["uploaded","processing","processed","failed"],"title":"DocumentStatus"},"DocumentUploadResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"filename":{"type":"string","title":"Filename"},"status":{"$ref":"#/components/schemas/DocumentStatus"}},"type":"object","required":["id","filename","status"],"title":"DocumentUploadResponse","description":"Returned immediately by POST /api/documents/upload."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"app_name":{"type":"string","title":"App Name"},"environment":{"type":"string","title":"Environment"},"database":{"$ref":"#/components/schemas/DatabaseStatus"}},"type":"object","required":["status","app_name","environment","database"],"title":"HealthResponse"},"RetrievalSearchRequest":{"properties":{"document_id":{"type":"string","format":"uuid","title":"Document Id"},"query":{"type":"string","maxLength":2000,"minLength":1,"title":"Query"}},"type":"object","required":["document_id","query"],"title":"RetrievalSearchRequest"},"RetrievalSearchResponse":{"properties":{"query":{"type":"string","title":"Query"},"document_id":{"type":"string","format":"uuid","title":"Document Id"},"vector_weight":{"type":"number","title":"Vector Weight"},"keyword_weight":{"type":"number","title":"Keyword Weight"},"results":{"items":{"$ref":"#/components/schemas/RetrievedChunk"},"type":"array","title":"Results"}},"type":"object","required":["query","document_id","vector_weight","keyword_weight","results"],"title":"RetrievalSearchResponse","description":"Includes the query and weighting configuration alongside the\nresults, so a response is self-contained for debugging/demoing\nwithout needing to cross-reference server config."},"RetrievedChunk":{"properties":{"chunk_id":{"type":"string","format":"uuid","title":"Chunk Id"},"page_number":{"type":"integer","title":"Page Number"},"section":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Section"},"content":{"type":"string","title":"Content"},"vector_score":{"type":"number","title":"Vector Score"},"keyword_score":{"type":"number","title":"Keyword Score"},"hybrid_score":{"type":"number","title":"Hybrid Score"}},"type":"object","required":["chunk_id","page_number","section","content","vector_score","keyword_score","hybrid_score"],"title":"RetrievedChunk"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}